requirements.txt
pinecone-client
tqdm
langchain
llama-cpp-python
python-dotenv
openai
tiktoken
streamlit
html2text
bs4
transformers
nltk